+++ tests/__init__.py
@@ -11,3 +11,4 @@ from os import path
 tests_path = path.abspath(path.dirname(__file__))
 
 test_server_path = path.join(tests_path, "test_server.py")
+sample_daemon_path = path.join(tests_path, "sample_daemon.py")