+++ mirakuru/__init__.py
@@ -31,7 +31,7 @@ from mirakuru.exceptions import (
     AlreadyRunning,
 )
 
-__version__ = '0.4.0'
+__version__ = '0.5.0'
 
 __all__ = (
     'Executor',