+++ CHANGES.rst
@@ -1,6 +1,11 @@
 CHANGELOG
 =========
 
+unreleased
+-------
+
+* Corrected code to conform with W503, D210 and E402 linters errors as reported by pylama 6.3.1
+
 0.4.0
 -------