+++ CHANGES.rst
@@ -5,8 +5,8 @@ unreleased
 ----------
 
 - Corrected code to conform with W503, D210 and E402 linters errors as reported by pylama 6.3.1
-- [feature] introduces a hack that kills subprocess of executor process
-  requires 'ps xe -ww' command being available in OS otherwise logs error
+- [feature] introduces a hack that kills all subprocesses of executor process.
+  It requires 'ps xe -ww' command being available in OS otherwise logs error.
 
 0.4.0
 -------