+++ CHANGES.rst
@@ -7,7 +7,7 @@ unreleased
 - Corrected code to conform with W503, D210 and E402 linters errors as reported by pylama 6.3.1
 - [feature] introduces a hack that kills all subprocesses of executor process.
   It requires 'ps xe -ww' command being available in OS otherwise logs error.
-- [refactoring] Classes name convetion change.
+- [refactoring] Classes name convention change.
   Executor class got renamed into SimpleExecutor and StartCheckExecutor class got renamed into Executor.
 
 0.4.0