+++ CHANGES.rst
@@ -1,7 +1,7 @@
 CHANGELOG
 =========
 
-unreleased
+0.5.0
 ----------
 
 - Corrected code to conform with W503, D210 and E402 linters errors as reported by pylama 6.3.1